requirements.txt
kafka-python==2.0.2
boto3==1.34.0